Eliminar el periodo de la base de datos luego de ejecutarse el test.

+++ src/test/java/test/PersistenciaTest.java
@@ -33,33 +33,31 @@ public class PersistenciaTest
 		  LocalDate fechaFin = LocalDate.parse("2008-01-01");
 		  Periodo periodo = new Periodo(fechaInicio, fechaFin);		  
 		 
-		  String obtenerPeriodoGrabado = "FROM Periodo";
+		  String obtenerPeriodosGrabados = "FROM Periodo";
 		  		  
 		  em.getTransaction().begin();
 		  em.persist(periodo);
 		  em.flush();
 		  em.getTransaction().commit();
 
-		  List<Periodo> periodos = em.createQuery(obtenerPeriodoGrabado, Periodo.class).getResultList();
+		  List<Periodo> periodos = em.createQuery(obtenerPeriodosGrabados, Periodo.class).getResultList();
 		  
-		  Assert.assertTrue(periodos.size() >= 0);
+		  Assert.assertTrue(periodos.size() > 0);
 	  }
 	  
 	  @After
 	  public void limpiar() 
-	  {
-		
-		  System.out.println("------------->");
+	  {		  
+		  String obtenerPeriodoGrabado = "FROM Periodo WHERE fechaInicio='2007-01-01' AND fechaFin='2008-01-01'";		  
+		  List<Periodo> periodoList = em.createQuery(obtenerPeriodoGrabado, Periodo.class).getResultList();
 		  
-		  LocalDate fechaInicio = LocalDate.parse("2007-01-01");
-		  LocalDate fechaFin = LocalDate.parse("2008-01-01");
-		  Periodo periodo = new Periodo(fechaInicio, fechaFin);		  
-
-		  Periodo periodoElem = (Periodo) em.find(Periodo.class,  periodo);
-		  
-		  em.getTransaction().begin();
-		  em.remove(periodoElem);
-		  em.getTransaction().commit();  
-		  
+		  if(periodoList.size() > 0)
+		  {			  						  			  			  
+			  Periodo periodoElem = (Periodo) periodoList.get(0);
+			  em.getTransaction().begin();
+			  em.remove(periodoElem);
+			  em.getTransaction().commit();  			  
+		  }
+			  
 	  }
 }
